+++ services/web/test/API/Shows/Slug/Episode/Get/HandlerSpec.hs
@@ -51,8 +51,8 @@ mkEpisodeInsert showId templateId userId =
       eiAudioMimeType = Nothing,
       eiDurationSeconds = Nothing,
       eiArtworkUrl = Nothing,
-      eiScheduleTemplateId = templateId,
-      eiScheduledAt = read "2020-01-01 10:00:00 UTC",
+      eiScheduleTemplateId = Just templateId,
+      eiScheduledAt = Just (read "2020-01-01 10:00:00 UTC"),
       eiCreatedBy = userId
     }